requirements.txt
click
pandas
omegaconf
tqdm
wandb
lmdb
openai
anthropic
Pillow
seaborn
scipy
scikit-learn
fuzzywuzzy
plotly
sentence_transformers
torch
clip
torchvision
InstructorEmbedding
open_clip_torch
flask